@@ -13,10 +13,10 @@ classifier =
         Intended Audience :: Science/Research
 	License :: OSI Approved :: Apache Software License
         Operating System :: POSIX
-	Programming Language :: Python :: 3.6
 	Programming Language :: Python :: 3.7
 	Programming Language :: Python :: 3.8
 	Programming Language :: Python :: 3.9
+	Programming Language :: Python :: 3.10
 keywords =
     ocr
     htr
@@ -34,7 +34,7 @@ max_line_length = 160
 exclude = tests/*
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.7,<=3.10
 install_requires =
     jsonschema
     lxml
@@ -48,10 +48,12 @@ install_requires =
     jinja2
     python-bidi
     torchvision>=0.5.0
-    torch>=1.5.0,<1.11
+    torch>=1.7.1,<1.11
     coremltools>=3.3
     scikit-image>=0.17.0,<=0.19.1
     shapely
+    pyarrow
+    pytorch-lightning
 
 [options.extras_require]
 test = hocr-spec; pytest
