@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
-import os
 import unittest
 
 from pytest import raises
 
 from PIL import Image
+from pathlib import Path
 from kraken.binarization import nlbin
 
 from kraken.lib.exceptions import KrakenInputException
 
-thisfile = os.path.abspath(os.path.dirname(__file__))
-resources = os.path.abspath(os.path.join(thisfile, 'resources'))
+thisfile = Path(__file__).resolve().parent
+resources = thisfile / 'resources'
 
 class TestBinarization(unittest.TestCase):
 
@@ -29,14 +29,14 @@ class TestBinarization(unittest.TestCase):
         """
         Test that mode '1' images aren't binarized again.
         """
-        with Image.open(os.path.join(resources, 'bw.png')) as im:
+        with Image.open(resources / 'bw.png') as im:
             self.assertEqual(im, nlbin(im))
 
     def test_binarize_no_bw(self):
         """
         Tests binarization of image formats without a 1bpp mode (JPG).
         """
-        with Image.open(os.path.join(resources, 'input.jpg')) as im:
+        with Image.open(resources / 'input.jpg') as im:
             res = nlbin(im)
             # calculate histogram and check if only pixels of value 0/255 exist
             self.assertEqual(254, res.histogram().count(0), msg='Output not '
@@ -46,7 +46,7 @@ class TestBinarization(unittest.TestCase):
         """
         Tests binarization of RGB TIFF images.
         """
-        with Image.open(os.path.join(resources, 'input.tif')) as im:
+        with Image.open(resources /'input.tif') as im:
             res = nlbin(im)
             # calculate histogram and check if only pixels of value 0/255 exist
             self.assertEqual(254, res.histogram().count(0), msg='Output not '
@@ -56,7 +56,7 @@ class TestBinarization(unittest.TestCase):
         """
         Test binarization of mode 'L' images.
         """
-        with Image.open(os.path.join(resources, 'input.tif')) as im:
+        with Image.open(resources / 'input.tif') as im:
             res = nlbin(im.convert('L'))
             # calculate histogram and check if only pixels of value 0/255 exist
             self.assertEqual(254, res.histogram().count(0), msg='Output not '
