@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 import unittest
-import os
 
 from PIL import Image
+from pathlib import Path
 from pytest import raises
 
 from kraken.lib import lineest
 
-thisfile = os.path.abspath(os.path.dirname(__file__))
-resources = os.path.abspath(os.path.join(thisfile, 'resources'))
+thisfile = Path(__file__).resolve().parent
+resources = thisfile / 'resources'
 
 class TestLineest(unittest.TestCase):
 
@@ -23,7 +23,7 @@ class TestLineest(unittest.TestCase):
         """
         Test dewarping of a single line in B/W
         """
-        with Image.open(os.path.join(resources, '000236.png')) as im:
+        with Image.open(resources / '000236.png') as im:
             o = lineest.dewarp(self.lnorm, im.convert('1'))
             self.assertEqual(self.lnorm.target_height, o.size[1])
 
@@ -31,7 +31,7 @@ class TestLineest(unittest.TestCase):
         """
         Test dewarping of a single line in grayscale
         """
-        with Image.open(os.path.join(resources, '000236.png')) as im:
+        with Image.open(resources /'000236.png') as im:
             o = lineest.dewarp(self.lnorm, im.convert('L'))
             self.assertEqual(self.lnorm.target_height, o.size[1])
 
@@ -40,14 +40,14 @@ class TestLineest(unittest.TestCase):
         Test dewarping of a color line fails
         """
         with raises(ValueError):
-            with Image.open(os.path.join(resources, '000236.png')) as im:
+            with Image.open(resources /'000236.png') as im:
                 lineest.dewarp(self.lnorm, im.convert('RGB'))
 
     def test_dewarp_bw_undewarpable(self):
         """
         Test dewarping of an undewarpable line.
         """
-        with Image.open(os.path.join(resources, 'ONB_ibn_19110701_010.tif_line_1548924556947_449.png')) as im:
+        with Image.open(resources /'ONB_ibn_19110701_010.tif_line_1548924556947_449.png') as im:
             o = lineest.dewarp(self.lnorm, im)
             self.assertEqual(self.lnorm.target_height, o.size[1])
 
