@@ -13,7 +13,7 @@ from coremltools.proto import NeuralNetwork_pb2
 # all tensors are ordered NCHW, the "feature" dimension is C, so the output of
 # an LSTM will be put into C same as the filters of a CNN.
 
-__all__ = ['MaxPool', 'Reshape', 'Dropout', 'TransposedSummarizingRNN', 'LinSoftmax', 'ActConv2D']
+__all__ = ['Addition', 'MaxPool', 'Reshape', 'Dropout', 'TransposedSummarizingRNN', 'LinSoftmax', 'ActConv2D']
 
 
 class MultiParamSequential(Sequential):
@@ -28,6 +28,22 @@ class MultiParamSequential(Sequential):
             else:
                 inputs = module(inputs)
         return inputs
+
+
+class MultiParamParallel(Module):
+    """
+    Parallel module.
+    """
+    def forward(self, *inputs):
+        outputs = []
+        seq_lens = None
+        for module in self._modules.values():
+            if type(inputs) == tuple:
+                output, seq_lens = module(*inputs)
+                outputs.append(output)
+            else:
+                outputs.append(module(inputs))
+        return torch.cat(outputs, dim=1), seq_lens
 
 
 def PeepholeLSTMCell(input: torch.Tensor,
@@ -142,6 +158,97 @@ class PeepholeBidiLSTM(Module):
     @property
     def all_weights(self):
         return [[getattr(self, weight) for weight in weights] for weights in self._all_weights]
+
+
+class Addition(Module):
+    """
+    An addition module
+    """
+    def __init__(self, dim: int, chunk_size: int) -> None:
+        """
+        An addition module
+
+        Shape:
+             - Inputs: :math:`(N, C, H, W)` where `N` batches, `C` channels, `H`
+              height, and `W` width.
+            - Outputs output :math:`(N, C, H, W)`
+        """
+        self.dim = dim
+        self.chunk_size = chunk_size
+        super().__init__()
+
+    def forward(self, inputs: torch.Tensor, seq_len: Optional[torch.Tensor] = None) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:
+        out = inputs.unfold(self.dim, self.chunk_size, self.chunk_size)
+        out = out.sum(self.dim, keepdim=True)
+        out = out.transpose(-1, self.dim).squeeze(-1)
+        return out, seq_len
+
+    def get_shape(self, input: Tuple[int, int, int, int]) -> Tuple[int, int, int, int]:
+        """
+        Calculates the output shape from input 4D tuple NCHW.
+        """
+        input = list(input)
+        input[self.dim] = self.chunk_size
+        self.output_shape = tuple(input)
+        return self.output_shape
+
+    def deserialize(self, name, spec):
+        """
+        Noop for deserialization
+        """
+        pass
+
+    def serialize(self, name, input, builder):
+        params = NeuralNetwork_pb2.CustomLayerParams()
+        params.className = 'addition'
+        params.description = 'An addition layer'
+        params.parameters['dim'].intValue = self.dim
+        params.parameters['chunk_size'].intValue = self.chunk_size
+
+        builder.add_custom(name,
+                           input_names=[input],
+                           output_names=[name],
+                           custom_proto_spec=params)
+        return name
+
+
+class Identity(Module):
+    """
+    A placeholder identity operator.
+    """
+    def __init__(self) -> None:
+        """
+        A placeholder identity operator (mostly used for residual connections and similar).
+
+        Shape:
+             - Inputs: :math:`(N, C, H, W)` where `N` batches, `C` channels, `H`
+              height, and `W` width.
+            - Outputs output :math:`(N, C, H, W)`
+        """
+        super().__init__()
+
+    def forward(self, inputs: torch.Tensor, seq_len: Optional[torch.Tensor] = None) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:
+        return inputs, seq_len
+
+    def get_shape(self, input: Tuple[int, int, int, int]) -> Tuple[int, int, int, int]:
+        self.output_shape = input
+        return input
+
+    def deserialize(self, name, spec):
+        """
+        Noop for deserialization
+        """
+        pass
+
+    def serialize(self, name, input, builder):
+        params = NeuralNetwork_pb2.CustomLayerParams()
+        params.className = 'identity'
+        params.description = 'An identity layer'
+        builder.add_custom(name,
+                           input_names=[input],
+                           output_names=[name],
+                           custom_proto_spec=params)
+        return name
 
 
 class Reshape(Module):
